# docs/api/evalsmith-public.openapi.yaml
openapi: 3.1.0
info:
  title: EvalSmith Public API
  version: 0.1.0
  summary: Gateway-facing API surface for EvalSmith trial and private deployments
  description: >
    First-pass public contract for EvalSmith. This spec focuses on the core
    integration paths external users need for auth, datasets, traces,
    experiments, annotation, and monitoring.
servers:
  - url: http://127.0.0.1:8080
    description: Local gateway
...
paths:
...
  /api/v1/datasets:
    get:
      tags: [datasets]
      operationId: listDatasets
      summary: List datasets for the active project
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
        - name: name
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      security:
        - sessionCookie: []
        - bearerApiKey: []
      responses:
        "200":
          description: Dataset page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDatasetEnvelope"
    post:
      tags: [datasets]
      operationId: createDataset
      summary: Create a dataset
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
      security:
        - sessionCookie: []
        - bearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDatasetRequest"
      responses:
        "200":
          description: Dataset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetEnvelope"
  /api/v1/datasets/{dataset_id}/examples:
    get:
      tags: [datasets]
      operationId: listDatasetExamples
      summary: List examples in a dataset
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
        - $ref: "#/components/parameters/DatasetIdPath"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: version
          in: query
          schema:
            type: integer
            minimum: 1
          description: Read examples from a historical dataset snapshot
      security:
        - sessionCookie: []
        - bearerApiKey: []
      responses:
        "200":
          description: Example page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedGenericEnvelope"
...
components:
  securitySchemes:
    bearerApiKey:
      type: http
      scheme: bearer
      __REDACTED_SECRET__: API Key
    sessionCookie:
      type: apiKey
      in: cookie
      name: evalsmith_session
  parameters:
    ProjectIdHeader:
      name: X-Project-ID
      in: header
      required: true
      schema:
        type: string
      description: Active project identifier
    DatasetIdPath:
      name: dataset_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    EnvelopeBase:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        details:
          description: Optional error details
      required: [code, message]
...
    Dataset:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        current_version:
          type: integer
        example_count:
          type: integer
      required: [id, name]
    DatasetEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/Dataset"
    PaginatedDatasetEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/Dataset"
                total:
                  type: integer
                page:
                  type: integer
                page_size:
                  type: integer
              required: [items, total, page, page_size]
    PaginatedGenericEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items: {}
                total:
                  type: integer
                page:
                  type: integer
                page_size:
                  type: integer
                total_pages:
                  type: integer
    CreateDatasetRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        schema_def:
          type: object
      required: [name]
...